requirement.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-phonenumber-field==8.1.0
django-tailwind==4.0.1
djangorestframework==3.16.0
idna==3.10
phonenumbers==9.0.8
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0

